requirements.txt
pydantic
langgraph 
langsmith
langchain 
langchain_groq 
langchain_community
python-dotenv
arvix
wikipedia
gradio